fix: Fix importers in browsers
